requirements.txt
fastapi
requests
psutil
pyjwt
gunicorn
uvicorn[standard]
python-dotenv